move federal and state waste codes under the 'waste' resource hierarchy in our API
