+++ server/apps/trak/models/__init__.py
@@ -2,10 +2,9 @@
 Model definitions for the trak domain
 """
 
-from .code_models import WasteCode
 from .contact_models import ManifestPhone
 from .handler_models import Handler
 from .manifest_models import AdditionalInfo, Manifest
 from .signature_models import ESignature, PaperSignature, QuickerSign, Signer
 from .transporter_models import Transporter
-from .waste_models import WasteLine
+from .waste_models import WasteCode, WasteLine
