+++ server/apps/trak/urls.py
@@ -23,9 +23,16 @@ urlpatterns = [
                 # MTN
                 path("mtn", MtnList.as_view()),
                 path("mtn/<str:epa_id>", MtnList.as_view()),
-                # Codes
-                path("code/waste/federal", FederalWasteCodesView.as_view()),
-                path("code/waste/state/<str:state_id>", StateWasteCodesView.as_view()),
+                # waste info
+                path(
+                    "waste/",
+                    include(
+                        [
+                            path("code/federal", FederalWasteCodesView.as_view()),
+                            path("code/state/<str:state_id>", StateWasteCodesView.as_view()),
+                        ]
+                    ),
+                ),
             ]
         ),
     ),
